requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# C:\Users\1998a\Documents\GitHub\registro_trades\drive_test.ipynb: 1,2
# C:\Users\1998a\Documents\GitHub\registro_trades\quickstart.py: 1,6
PyDrive == 1.3.1

# C:\Users\1998a\Documents\GitHub\registro_trades\streamlit_app.py: 8
numpy == 1.19.5

# C:\Users\1998a\Documents\GitHub\registro_trades\streamlit_app.py: 3
omelette == 1.1.6

# C:\Users\1998a\Documents\GitHub\registro_trades\streamlit_app.py: 2
pandas == 1.2.0

# C:\Users\1998a\Documents\GitHub\registro_trades\streamlit_app.py: 3
pillow == 8.2.0

# C:\Users\1998a\Documents\GitHub\registro_trades\streamlit_app.py: 3
pillow-simd == 7.0.0.post3

# C:\Users\1998a\Documents\GitHub\registro_trades\streamlit_app.py: 1
streamlit == 0.82.0

# C:\Users\1998a\Documents\GitHub\registro_trades\streamlit_app.py: 7
xlrd == 2.0.1
